requirements.txt
torch==1.11.0+cu113
torchvision==0.12.0+cu113
numpy==1.23.5
chumpy==0.70
scipy==1.8.1
scikit-image==0.19.2
scikit-learn==1.1.1
opencv-python==4.6.0.66
Pillow==9.2.0
termcolor==1.1.0
PyYAML==6.0
tqdm==4.64.0
absl-py==1.1.0
tensorboard==2.11.0
tensorboardX==2.5
PyOpenGL==3.1.0
pyrender==0.1.45
trimesh==3.12.5
click==8.1.7
omegaconf==2.3.0
segmentation_models_pytorch==0.3.3
timm==0.9.2
psutil==5.8.0
lmdb==1.2.1
einops==0.3.2
kornia==0.6.12
gdown
plyfile
zmq